src/main/java/StudentService.java
import java.util.List;
import java.util.Optional;
import java.util.UUID;

public class StudentService {

    private final StudentRepo repo = new StudentRepo();

    public Student addNewStudent(Student newStudent) {
        Student studentToSave = newStudent.withId(UUID.randomUUID().toString());
        return repo.save(studentToSave);
    }

    public List<Student> getAllStudents(){
        return repo.getAllStudents();
    }

    public Student getById(String id) throws IdNotFoundException {
        Optional<Student> found= repo.findStudentById(id);
        return found.orElseThrow(IdNotFoundException::new);
    }
}
